simplify AtCoderClient.py
